# Refactoring: shift between Labels and LabelAdapters to align better with dedupelabels