--- a/pkg/ingester/client/compat_test.go
+++ b/pkg/ingester/client/compat_test.go
@@ -7,10 +7,8 @@ package client
 
 import (
 	"reflect"
-	"strconv"
 	"testing"
 
-	"github.com/grafana/mimir/pkg/mimirpb"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/assert"
@@ -86,54 +84,4 @@ func TestLabelNamesRequest(t *testing.T) {
 	assert.Equal(t, int64(mint), actualMinT)
 	assert.Equal(t, int64(maxt), actualMaxT)
 	assert.Equal(t, matchers, actualMatchers)
-}
-
-// The main usecase for `LabelsToKeyString` is to generate hashKeys
-// for maps. We are benchmarking that here.
-func BenchmarkSeriesMap(b *testing.B) {
-	benchmarkSeriesMap(100000, b)
-}
-
-func benchmarkSeriesMap(numSeries int, b *testing.B) {
-	series := makeSeries(numSeries)
-	sm := make(map[string]int, numSeries)
-
-	b.ReportAllocs()
-	b.ResetTimer()
-	for n := 0; n < b.N; n++ {
-		for i, s := range series {
-			sm[LabelsToKeyString(mimirpb.FromLabelAdaptersToLabels(s))] = i
-		}
-
-		for _, s := range series {
-			_, ok := sm[LabelsToKeyString(mimirpb.FromLabelAdaptersToLabels(s))]
-			if !ok {
-				b.Fatal("element missing")
-			}
-		}
-
-		if len(sm) != numSeries {
-			b.Fatal("the number of series expected:", numSeries, "got:", len(sm))
-		}
-	}
-}
-
-func makeSeries(n int) [][]mimirpb.LabelAdapter {
-	series := make([][]mimirpb.LabelAdapter, 0, n)
-	for i := 0; i < n; i++ {
-		series = append(series, []mimirpb.LabelAdapter{
-			{Name: "label0", Value: "value0"},
-			{Name: "label1", Value: "value1"},
-			{Name: "label2", Value: "value2"},
-			{Name: "label3", Value: "value3"},
-			{Name: "label4", Value: "value4"},
-			{Name: "label5", Value: "value5"},
-			{Name: "label6", Value: "value6"},
-			{Name: "label7", Value: "value7"},
-			{Name: "label8", Value: "value8"},
-			{Name: "label9", Value: strconv.Itoa(i)},
-		})
-	}
-
-	return series
 }