--- a/pkg/mimirpb/compat.go
+++ b/pkg/mimirpb/compat.go
@@ -104,6 +104,19 @@ func FromLabelAdaptersToMetric(ls []LabelAdapter) model.Metric {
 	return util.LabelsToMetric(FromLabelAdaptersToLabels(ls))
 }
 
+// FromLabelAdaptersToKeyString makes a string to be used as a key to a map.
+// It's much simpler than FromLabelAdaptersToString, but not human-readable.
+func FromLabelAdaptersToKeyString(ls []LabelAdapter) string {
+	buf := make([]byte, 0, 1024)
+	for i := range ls {
+		buf = append(buf, '\xff')
+		buf = append(buf, ls[i].Name...)
+		buf = append(buf, '\xff')
+		buf = append(buf, ls[i].Value...)
+	}
+	return string(buf)
+}
+
 // FromLabelAdaptersToString formats label adapters as a metric name with labels, while preserving
 // label order, and keeping duplicates. If there are multiple "__name__" labels, only
 // first one is used as metric name, other ones will be included as regular labels.