--- a/pkg/util/labels.go
+++ b/pkg/util/labels.go
@@ -8,19 +8,8 @@ package util
 import (
 	"strings"
 
-	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/labels"
 )
-
-// LabelsToMetric converts a Labels to Metric
-// Don't do this on any performance sensitive paths.
-func LabelsToMetric(ls labels.Labels) model.Metric {
-	m := make(model.Metric, ls.Len())
-	ls.Range(func(l labels.Label) {
-		m[model.LabelName(l.Name)] = model.LabelValue(l.Value)
-	})
-	return m
-}
 
 // LabelMatchersToString returns a string representing the input label matchers.
 func LabelMatchersToString(matchers []*labels.Matcher) string {