#### What this PR does

This improves efficiency, mostly in tests but a couple of non-tests too.

Split into four commits:
* storegateway tests: re-use Builder when creating labels - this is more efficient, especially with dedupelabels.
* storegateway: retain Builder in loadingSeriesChunkRefsSetIterator - saves memory allocation when moving to next series.
* mimirpb: add FromLabelAdaptersToKeyString - avoid converting to labels and then to string.
* labels: avoid double-converting around FromLabelAdaptersToMetric 

Benchmarks (after changing the benchmark to match usage):
```
             │ before.txt  │             after.txt              │
             │   sec/op    │   sec/op     vs base               │
SeriesMap-28   46.13m ± 2%   22.93m ± 5%  -50.29% (p=0.002 n=6)

             │  before.txt   │              after.txt              │
             │     B/op      │     B/op      vs base               │
SeriesMap-28   115.98Mi ± 0%   27.47Mi ± 0%  -76.31% (p=0.002 n=6)

             │ before.txt  │             after.txt              │
             │  allocs/op  │  allocs/op   vs base               │
SeriesMap-28   600.1k ± 0%   200.0k ± 0%  -66.67% (p=0.002 n=6)
```

#### Checklist

- [x] Tests updated.
- NA Documentation added.
- NA `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- NA [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.